refactor: make a function for formatting request parameters for a single endpoint
